bot-pilot-chat/backend/chat-backend/requirements.txt
fastapi
uvicorn
python-dotenv
sentence-transformers
tensorflow[and-cuda]==2.12.0
httpx
pytest
pytest-cov
rasa==3.6.21
spacy==3.7.4
